Add support for MergeMapper, with LoadOrder Form lookup now used as a proxy to TESDataHandler Move Missives Board blacklist to a Builtin Collection, disabled if mod not present of load order so dictates Resolve TESObjectCONT into Collections Fix compiler warning error

+++ src/FormHelpers/IHasValueWeight.cpp
@@ -57,7 +57,8 @@ bool IHasValueWeight::ValueWeightTooLowToLoot() const
 	double weight = GetWeight();
 	if (weight == 0.0)
 	{
-		bool result(SettingsCache::Instance().CheckWeightlessValue() && worth < SettingsCache::Instance().WeightlessMinimumValue());
+		bool result(SettingsCache::Instance().CheckWeightlessValue() &&
+			worth < static_cast<uint32_t>(SettingsCache::Instance().WeightlessMinimumValue()));
 		DBG_VMESSAGE("Weightless item value too low to loot? {}", worth, result);
 		return result;
 	}
